main.go: fix OIDC
This commit fixes the OIDC configuration so that the thanos-rule-syncer
can properly authenticate with the Observatorium API. Previsouly, the
OIDC configuration was never applied to the HTTP client's underlying
transport, meaning that it had no effect and all requests would fail.

This commit also makes the interval at which the Observatorium API is
polled configurable.

Finally, this commit fixes the thanos-rule-syncer so that it no longer
uses hard-coded paths for the Observatorium CA.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>