test/integration.sh: retry query
Currently, querying for the time series produced by the ruler is racey
because the ruler may not yet have had a chance to generate the time
series specified by the recording rules. This raciness can cause the
integration test to occasionally fail. This commit adds retries to the
testing of the rules. Eventually, these integration tests could be
ported to the same e2e test framework used in observatorium/api.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>